requirements.txt
fastapi
uvicorn
openai
firebase-admin
python-dotenv
python-multipart
pillow
pytesseract
numpy
